Change player_score type from Float32 to Float64 for improved precision

+++ migrations/clickhouse/04_create_mmr_history_table.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS mmr_history
  algorithm     Enum8('LinearRegression' = 0) default 'LinearRegression',
  account_id    UInt32,
  match_id      UInt64,
- player_score  Float32,
+ player_score  Float64,
  rank          UInt32 ALIAS [
   0,
   11,
@@ -89,7 +89,7 @@ CREATE TABLE IF NOT EXISTS hero_mmr_history
  account_id    UInt32,
  hero_id       UInt8,
  match_id      UInt64,
- player_score  Float32,
+ player_score  Float64,
  rank          UInt32 ALIAS [
   0,
   11,
